more refactor - attach to console devtool helper.
