+++ src/3d/camera/dialogueCamera.ts
@@ -1,6 +1,6 @@
 import {  Orientation } from "@/shared/types/3d.types";
 import { CameraController } from "./createCameraController";
-import { XYZ } from "./PlayerCam";
+import { XYZ } from "@/shared/types/3d.types";
 import { DialogueNode } from "@/core/dialogue/dialogueNode.types";
 import { DialogueService, StartDialogueOptions } from "@/core/dialogue/dialogueService";
