+++ src/3d/pipeline/applyShadows.ts
@@ -1,4 +1,4 @@
-import { DirectionalLight, Element3D, ObjModel } from "lume";
+import { Element3D, ObjModel } from "lume";
 
 
 /** Recursively enables shadows (including self-shadowing) on all subcomponents of model. Can be used to add shadows to a map component. */
@@ -30,8 +30,4 @@ export default function applyShadows(ref: Element3D | ObjModel) {
         });
     })
 
-}
-
-export function tightenShadows(dirLight: DirectionalLight, boundingObject: Element3D) {
-    // TODO - Restrict shadow camera to just be inside a bounding box set by the side of bounding object
 }
