+++ src/devtools/attachToConsole.ts
@@ -1,0 +1,3 @@
+export default function attachToConsole(thing: any, as: string) {
+    (window as any)[as] = thing;
+}
